initial working version
This creates a `@vscode/test-cli` NPM module that exposes a `vscode-test`
command line.

This reads a configuration file, described in the readme, and runs all
tests defined therein by default. Multiple configurations can be defined,
in which case they are all run sequentially.

See previous discussion on https://github.com/microsoft/vscode-test/pull/235
This is largely the same as the version in that PR, with the addition
of a reporter and addressed comments.