Fix Resolve in bin.mts (#10)
bin.mts is an ES module, but tries to use require.resolve which is only available to CommonJS modules. This generates a requires so things work properly. The future option is to use `import.meta.requires` if it ever goes stable. This type of resolution is also generally not necessary, passing through the relative paths works fine as well.